next_player() This function has been improved, now the players can go on the board based on the dice rolls. One player rolls, goes a few steps then another player comes.
Buying property also works, based on the player's money. The player's
money does not decrease however after a purchase. :( That needs to be
fixed.

The 'round complete' problem also needs to be addressed, as it is not
known yet what happens if  a player reaches the last field and goes
beyond. If it works, the player must get +200 money as well!